+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Version History
 
-## unreleased
+## Development
+
+### Changed
+
+* Registering of data items will now work for files, directories and symlinks created in
+the watch directory.
+* crc32c is now used during metadata generation to improve performance.
+
+## 0.1.0
 
 Added
 -----
 
 * Added Kafka server
 
-
 * Bootstrap repo with ska-cookiecutter-pypackage
+
+* Added directory watcher